move xlsxreader identity statics to dfhack lib
