+++ library/PluginStatics.cpp
@@ -1,0 +1,9 @@
+#include "PluginStatics.h"
+
+namespace DFHack {
+
+// xlsxreader statics
+DFHACK_EXPORT xlsx_file_handle_identity xlsx_file_handle::_identity;
+DFHACK_EXPORT xlsx_sheet_handle_identity xlsx_sheet_handle::_identity;
+
+}
