+++ plugins/xlsxreader.cpp
@@ -8,37 +8,11 @@
 #include "DataFuncs.h"
 #include "LuaTools.h"
 #include "PluginManager.h"
+#include "PluginStatics.h"
 
 using namespace DFHack;
 
 DFHACK_PLUGIN("xlsxreader");
-
-struct xlsx_file_handle_identity : public compound_identity {
-    xlsx_file_handle_identity()
-        :compound_identity(0, nullptr, nullptr, "xlsx_file_handle") {};
-    DFHack::identity_type type() override { return IDTYPE_OPAQUE; }
-};
-
-struct xlsx_sheet_handle_identity : public compound_identity {
-    xlsx_sheet_handle_identity()
-        :compound_identity(0, nullptr, nullptr, "xlsx_sheet_handle") {};
-    DFHack::identity_type type() override { return IDTYPE_OPAQUE; }
-};
-
-struct xlsx_file_handle {
-    const xlsxioreader handle;
-    xlsx_file_handle(xlsxioreader handle): handle(handle) {}
-    static xlsx_file_handle_identity _identity;
-};
-
-struct xlsx_sheet_handle {
-    const xlsxioreadersheet handle;
-    xlsx_sheet_handle(xlsxioreadersheet handle): handle(handle) {}
-    static xlsx_sheet_handle_identity _identity;
-};
-
-xlsx_file_handle_identity xlsx_file_handle::_identity;
-xlsx_sheet_handle_identity xlsx_sheet_handle::_identity;
 
 // returns NULL on error
 xlsx_file_handle* open_xlsx_file(std::string filename) {
